Remove C++-style I/O operator
We need to decide between C-style and C++-style I/O. Currently, the
former seems to be the preferred choice. It would probably be better
to define I/O operators outside of their respecive classes, and in a
separate header file, so that they can be included only where needed.