Move Dot() and Cross() outside of Vector3D class
This change forces the use of Dot(a,b) and Cross(a, b) instead of
a.Dot(b) and a.Cross(b), as the first notation is clearer. It also
disallows the use of different template parameter types in the call,
as code using it is at best a performance issue, but likely an error.