Add files equivalent to old global header, new backend test
Files in the core/ directory contain macros supposed to substitute
the ones from the old Global.h header file. The macros have been split
into several files for convenience, and suggestions are made for their
new namings.

A small test to the backend was added, which should be later improved to
cover all of the features we intend to have in a backend. The vector
backend automatically falls back into a scalar backend when Vc is not
available. Therefore, it is safe to write code using the vector backend,
even if it is not available. Compatibility interfaces with Vc may not be
complete yet, in which case the classes in the scalar backend will need
to be extended to provide the missing Vc features.