Fix __sincos() to sincos() redirection on Mac OS
The defines were interfering with other places where an actual sincos()
function was implemented, such as in Vc. Instead of using defines, this
fixes the bug by defining an inline function that redirects to __sincos().

It is now also not necessary to provide a separate sincos() in the Vc
backend, which was only hidden by the defines before, so we remove it.
(That is, Vc::sincos() was replaced with Vc::__sincos() before, and that
did not exist.)