Improve support for subproject use and builtins
Client projects may use VecCore as a subproject, i.e. by adding
it to their CMake build through an add_subdirectory call. They may
also use find_package to locate an external install.

To support both use cases transparently, create VecCore::VecCore
IMPORTED target alongside VecCore. The target is reconstructed in
the same way as is done in VecCoreConfig to configure the requested
backend. It is made GLOBAL as client projects will always include
VecCore in a subdirectory. Client projects can link to VecCore::VecCore
independent of using VecCore as a subproject or external package.

The canonical ALIAS target method is not used as this requires export
of the VecCore target. This exports any build time choise of backend
as a dependency of the imported target, making configuration/use more
difficult.

VecCore can support build of optional backends like UMESIMD and Vc.
Simplify the use of builtin UMESIMD by adding it as a subproject using
FetchContent rather than ExternalProject. This structure is cleaner and
less prone to ordering/dependency issues.

This latter change requires a bump in CMake version to 3.11 to provide
FetchContent. A bump to CMake 3.12 has been made as this provides a
greater range of features, especially awareness of C++20.